formats: objfile idomatic reader/writer
